requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
python-dotenv==1.0.0
pyodbc==5.0.1
SQLAlchemy==2.0.25 
logging==0.5.1